UI 클래스의 디렉토리 변경
