+++ app/src/main/java/org/slipp/masil/games/web/PlayController.java
@@ -3,8 +3,8 @@ package org.slipp.masil.games.web;
 import org.slipp.masil.games.application.HighLowApplicationService;
 import org.slipp.masil.games.domains.highrow.GuessHighLowNumber;
 import org.slipp.masil.games.domains.highrow.HighLowJudgement;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -13,8 +13,34 @@ import org.springframework.web.servlet.ModelAndView;
 @Controller
 public class PlayController {
 
-    @Autowired
-    HighLowApplicationService highLowApplicationService;
+    private final HighLowApplicationService highLowApplicationService;
+
+    public PlayController(HighLowApplicationService highLowApplicationService) {
+        this.highLowApplicationService = highLowApplicationService;
+    }
+
+    @GetMapping(value = "/games/{gameId}")
+    public ModelAndView beforeGameStart(@PathVariable String gameId) {
+        ModelAndView mv = new ModelAndView();
+        mv.setViewName("highlow/beforePlaying");
+        mv.addObject("gameId", gameId);
+        return mv;
+    }
+
+    @PostMapping(value = "/games/{gameId}")
+    public ModelAndView getUserName(
+            @PathVariable String gameId,
+            @RequestParam(value = "userName") String userName
+    ) {
+        Long contextId = highLowApplicationService.start(userName);
+        ModelAndView mv = new ModelAndView();
+        mv.setViewName("highlow/playing");
+        mv.addObject("contextId", contextId);
+        mv.addObject("userName", userName);
+        mv.addObject("gameId", gameId);
+
+        return mv;
+    }
 
     @PostMapping(value = "/games/{gameId}/turn")
     public ModelAndView playingTurn(
@@ -27,10 +53,11 @@ public class PlayController {
         HighLowJudgement play = highLowApplicationService.play(GuessHighLowNumber.of(contextId, guessNumber));
 
         ModelAndView mv = new ModelAndView();
-        mv.setViewName("playingHighLowGame");
+        mv.setViewName("highlow/playing");
         mv.addObject("result", play.name());
         mv.addObject("userName", userName);
         mv.addObject("contextId", contextId);
+        mv.addObject("gameId", gameId);
         return mv;
     }
 }
