+++ app/src/main/resources/templates/highlow/ranking.html
