+++ app/src/test/java/org/slipp/masil/games/web/GamesControllerTest.java
@@ -8,12 +8,10 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
 
 @WebMvcTest
-class IndexControllerTest {
+class GamesControllerTest {
 
     @Autowired
     MockMvc mockMvc;
@@ -26,17 +24,8 @@ class IndexControllerTest {
 
     @Test
     void indexView() throws Exception {
-        mockMvc.perform(get("/"))
-                .andExpect(view().name("index"))
+        mockMvc.perform(get("/games"))
+                .andExpect(view().name("gameLounge"))
                 .andReturn();
-    }
-
-
-    @Test
-    void startHighLowGameButton() throws Exception {
-        mockMvc.perform(get("/games/{gameId}", 1))
-                .andExpect(view().name("beforeHighLowGameView"))
-                .andDo(print())
-                .andExpect(model().attribute("gameId", "1"));
     }
 }
